# openapi.yaml
openapi: "3.0.3"
info:
  title: 股票基金投资论坛 API
  description: |
    股票基金投资论坛后端 RESTful API。
    支持用户认证、内容发布、社交互动、搜索聚合、管理运营等完整功能。
  version: "1.0.0"
  contact:
    name: 贺嘉轩
    url: https://github.com/kkk431/Stock-Fund-Investment-Forum
servers:
  - url: http://localhost:8000
    description: 本地开发环境
  - url: http://api.example.com
    description: 生产环境（待部署）
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: 登录后返回的 JWT Token，放入 Authorization 请求头
  schemas:
...
    ApiResponse:
      type: object
      properties:
        code:
          type: integer
          description: 业务状态码
          example: 200
        message:
          type: string
          description: 提示消息
          example: success
        data:
          oneOf:
            - type: object
            - type: array
            - type: "null"
          description: 响应数据
      required: [code, message]
    PaginatedResponse:
      type: object
      properties:
        code:
          type: integer
          example: 200
        message:
          type: string
          example: success
        data:
          type: object
          properties:
            items:
              type: array
              items:
                type: object
            total:
              type: integer
              example: 100
            page:
              type: integer
              example: 1
            size:
              type: integer
              example: 20
      required: [code, message, data]
...
    Post:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        content_summary:
          type: string
        author:
          type: object
          properties:
            id:
              type: integer
            nickname:
              type: string
            avatar_url:
              type: string
            auth_level:
              type: string
        category:
          type: object
          properties:
            id:
              type: integer
            name:
              type: string
        post_type:
          type: string
          enum: [normal, long_article, poll, moment]
        view_count:
          type: integer
        like_count:
          type: integer
        comment_count:
          type: integer
        is_elite:
          type: boolean
        tags:
          type: array
          items:
            type: string
        created_at:
          type: string
          format: date-time
    PostDetail:
      allOf:
        - $ref: "#/components/schemas/Post"
        - type: object
          properties:
            content:
              type: string
            attachments:
              type: array
              items:
                type: object
                properties:
                  file_name:
                    type: string
                  file_url:
                    type: string
                  file_size:
                    type: integer
            collect_count:
              type: integer
            share_count:
              type: integer
            is_liked:
              type: boolean
            is_collected:
              type: boolean
            updated_at:
              type: string
              format: date-time
    CreatePostRequest:
      type: object
      required: [category_id, title, content, post_type]
      properties:
        category_id:
          type: integer
        title:
          type: string
          maxLength: 120
          minLength: 1
        content:
          type: string
        post_type:
          type: string
          enum: [normal, long_article, poll, moment]
        tags:
          type: array
          items:
            type: string
        attachments:
          type: array
          items:
            type: object
            properties:
              file_name:
                type: string
              file_url:
                type: string
...
    VoteRequest:
      type: object
      required: [option_ids]
      properties:
        option_ids:
          type: array
          items:
            type: integer
          minItems: 1
    VoteResult:
      type: object
      properties:
        option_id:
          type: integer
        option_text:
          type: string
        votes:
          type: integer
        percentage:
          type: number
          format: float

    # ==================== 收藏 ====================
...
paths:
...
  /api/posts:
    get:
      tags:
        - 内容系统
      summary: 获取帖子列表
      operationId: listPosts
      parameters:
        - name: category_id
          in: query
          schema:
            type: integer
          description: 板块ID筛选
        - name: sort
          in: query
          schema:
            type: string
            enum: [latest, hot, elite]
          description: 排序方式
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: size
          in: query
          schema:
            type: integer
            default: 20
            maximum: 50
        - name: keyword
          in: query
          schema:
            type: string
          description: 关键词搜索
      responses:
        "200":
          description: 成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaginatedResponse"
    post:
      tags:
        - 内容系统
      summary: 创建帖子
      operationId: createPost
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePostRequest"
      responses:
        "201":
          description: 发布成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponse"
  /api/posts/{id}:
    get:
      tags:
        - 内容系统
      summary: 获取帖子详情
      operationId: getPostDetail
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: 成功
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/ApiResponse"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/PostDetail"
        "404":
          description: 帖子不存在
    put:
      tags:
        - 内容系统
      summary: 编辑帖子
      operationId: updatePost
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePostRequest"
      responses:
        "200":
          description: 编辑成功
    delete:
      tags:
        - 内容系统
      summary: 删除帖子
      operationId: deletePost
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: 删除成功
  /api/posts/{id}/vote:
    post:
      tags:
        - 内容系统
      summary: 参与投票
      operationId: votePost
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VoteRequest"
      responses:
        "200":
          description: 投票成功
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/ApiResponse"
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          results:
                            type: array
                            items:
                              $ref: "#/components/schemas/VoteResult"
...